backend/requirements.txt
flask 
flask-cors 
web3 
python-dotenv